+++ lib/prepare-query.js
@@ -1,5 +1,6 @@
 const _ = require('lodash')
 const sql = require('./alasql')
+const Query = require('./sql-query-builder')
 const {
   create: createSqlStatement,
   params: createSqlParams
