+++ lib/query.js
@@ -1,4 +1,5 @@
 const _ = require('lodash')
+const Query = require('./sql-query-builder')
 const {
   create: createSqlStatement
 } = require('./sql-query')
