+++ lib/sql-query-builder/index.js
@@ -1,5 +1,5 @@
 'use strict'
-const Geometry = require('./geometry')
+const Geometry = require('../geometry')
 const Where = require('./where')
 const Select = require('./select')
 const Order = require('./order')
