+++ lib/sql-query-builder/select/fields-fragment.js
