+++ lib/sql-query-builder/select/inline-functions-fragment.js
